feat(0046): 0 предупреждений сборки — Location u32×3 (−414 result_large_err), -D warnings
Свёл rustc 0 / clippy 0 в обоих крейтах и закрепил `-D warnings` на шаге clippy
в precheck.sh и CI (Option B ADR; CLI-уровень, не запрещённый deny(warnings)).

Находка: свежий инвентарь дал 549 clippy (не ~106 из снимка 2026-07-15), из них
414 — один класс result_large_err (включён по умолчанию в clippy 0.1.99;
Diagnostic = 136 байт > порога 128).

0046-01: Location::Source(u64,usize,usize) -> (u32,u32,u32). Вариант 16 байт
вместо 32 -> Diagnostic 120 байт -> все 414 result_large_err исчезли без
#[allow] и без правки сигнатур Result. Публичный API методов Location остался в
usize/String (каст в аксессорах + хелпер Location::source, им заменены 160
конструирований в grammar.lalrpop/лексере через sed; разбор — as usize у места).

0046-02: clippy --fix (needless_ref/collapsible_if/clone_on_copy/…) + ручные
#[allow] с обоснованием (too_many_arguments у печатников, large_enum_variant у
UnitKind, type_complexity, upper_case_acronyms, redundant_guards) + правки
(doc_lazy_continuation, field_reassign, assertions_on_constants,
needless_range_loop, get().is_none()) + 3 rustc.

0046-03: -D warnings на clippy в precheck.sh и новый шаг CI.

Вывод генераторов байт-в-байт неизменен (git diff examples/generated пуст).
precheck.sh EXIT=0; тесты 2099 passed; долг размеров 21->20. Крейт grammar
0.6.0 -> 0.7.0 (ломающая смена типов публичного Location::Source); язык не менялся.

+++ grammar/src/lsp/semantic_tokens.rs
@@ -194,7 +194,7 @@ pub fn semantic_tokens(source: &str) -> SemanticTokens {
             Comment::Line(loc, _) | Comment::DocLine(loc, _) | Comment::Block(loc, _) => loc,
         };
         if let Location::Source(_, start, end) = loc {
-            raw.push((*start, *end, TT_COMMENT));
+            raw.push((*start as usize, *end as usize, TT_COMMENT));
         }
     }
